refactor: extract memories client from router
