+++ backend/open_webui/utils/middleware.py
@@ -56,7 +56,7 @@ from open_webui.clients.pipelines import (
     process_pipeline_inlet_filter,
     process_pipeline_outlet_filter,
 )
-from open_webui.routers.memories import query_memory, QueryMemoryForm
+from open_webui.clients.memories import query_memory, QueryMemoryForm
 
 from open_webui.utils.webhook import post_webhook
 from open_webui.utils.files import (
